refactor CreateScheduleOptions and ScheduleInputs to accept fewer props
